+++ examples/game_of_life_async.lua
@@ -11,7 +11,7 @@ local print_every_iteration = false
 
 -- Domain and seed
 local sq = primitives.square(40,20)
-local rn = subpattern.random(sq, 0.5)
+local rn = subpattern.random(sq, math.floor(sq:size()*0.5))
 
 -- vonNeumann neighbourhood for pretty printing
 local nbh = neighbourhood.von_neumann()